requirements.txt
langchain
transformers
sentence-transformers
datasets
pytorch
torch
torchvision
torchaudio
faiss-cpu
langchain-community
transformers 
bitsandbytes 
accelerate
lxml
llama-index
llama-index-embeddings-huggingface
peft
auto-gptq
optimum
bitsandbytes
wheel
flash_attn
accelerate
requests
beautifulsoup4
flash_attn==2.5.8  
accelerate==0.31.0 
transformers==4.41.2 
beautifulsoup4
re
